Fjernet typer.ts og erstattet med egne filer for de ulike feltene
